حل التضارب وتحديث الكود

+++ Modules/Sport/Routes/api.php
@@ -17,10 +17,10 @@ use Modules\Sport\Http\Controllers\MatchSlotController;
 |
 */
 
-Route::middleware(['auth:api', 'role:admin'])->prefix('sport')->group(function () {
-    Route::post('/create', [SportController::class, 'create']);
-    Route::post('/update/{id}',[SportController::class,'update']);
-    Route::delete('/delete/{id}',[SportController::class,'delete']);
+Route::middleware(['auth:api'])->prefix('sport')->group(function () {
+    Route::post('/create', [SportController::class, 'create'])->middleware('role:admin');
+    Route::post('/update/{id}',[SportController::class,'update'])->middleware('role:admin');;
+    Route::delete('/delete/{id}',[SportController::class,'delete'])->middleware('role:admin');;
     Route::get('/view/{id}',[SportController::class,'view']);
     Route::get('/viewall',[SportController::class,'viewall']);
